Split in ui/ux
